Refactoring existing layout code into new eppic.assembly.layout package
